kcf use float data type rather than double.
In our practice, float is good enough and could get better performance.
With this patch, one of my benchmark could get about 20% performance gain.

+++ modules/tracking/src/precomp.hpp
@@ -50,7 +50,7 @@
 
 namespace cv
 {
-	extern const double ColorNames[][10];
+	extern const float ColorNames[][10];
 
     namespace tracking {
